refactor: added component for assets list
